Add collision controller

+++ BrickGame/core/src/com/mictlan/brick/entities/Brick.java
@@ -9,7 +9,7 @@ public class Brick extends GameObject {
     private int points = 100;
     private boolean hasPowerUp = false;
 
-    public Brick(int x, int y, int width, int height, boolean hasPowerUp) {
+    public Brick(float x, float y, float width, float height, boolean hasPowerUp) {
         super(x, y, width, height);
         this.hasPowerUp = hasPowerUp;
         color = ColorFactory.getColor(244, 244, 244);
